Move backoff to retry.go, used for upload/import retries
